# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Wishlist API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/public/w/{share_slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Wishlist",
        "operationId": "get_public_wishlist_api_v1_public_w__share_slug__get",
        "parameters": [
          {
            "name": "share_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Slug"
            }
          },
          {
            "name": "X-Guest-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Token"
            }
          },
          {
            "name": "wishlist_access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wishlist Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicWishlistView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/w/{share_slug}/guest-session": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Create Guest Session",
        "operationId": "create_guest_session_api_v1_public_w__share_slug__guest_session_post",
        "parameters": [
          {
            "name": "share_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestSessionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/w/{share_slug}/items/{item_id}/reserve": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Reserve Item",
        "operationId": "reserve_item_api_v1_public_w__share_slug__items__item_id__reserve_post",
        "parameters": [
          {
            "name": "share_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Slug"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Guest-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "public"
        ],
        "summary": "Unreserve Item",
        "operationId": "unreserve_item_api_v1_public_w__share_slug__items__item_id__reserve_delete",
        "parameters": [
          {
            "name": "share_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Slug"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Guest-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/w/{share_slug}/items/{item_id}/contributions": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Contribute To Item",
        "operationId": "contribute_to_item_api_v1_public_w__share_slug__items__item_id__contributions_post",
        "parameters": [
          {
            "name": "share_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Slug"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "X-Guest-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/w/{share_slug}/events": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Events",
        "operationId": "get_events_api_v1_public_w__share_slug__events_get",
        "parameters": [
          {
            "name": "share_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Slug"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContributionRequest": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "ContributionRequest"
      },
      "ContributionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "accepted_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Accepted Amount"
          },
          "collected_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Collected Amount"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent"
          }
        },
        "type": "object",
        "required": [
          "message",
          "item_id",
          "accepted_amount",
          "collected_amount",
          "progress_percent"
        ],
        "title": "ContributionResponse"
      },
...
      "EventsResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/RealtimeEventView"
            },
            "type": "array",
            "title": "Events"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "events",
          "next_cursor"
        ],
        "title": "EventsResponse"
      },
      "GuestItemView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "product_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "mode": {
            "$ref": "#/components/schemas/ItemMode"
          },
          "target_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount"
          },
          "collected_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Collected Amount"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "is_reserved": {
            "type": "boolean",
            "title": "Is Reserved"
          },
          "reserved_by_you": {
            "type": "boolean",
            "title": "Reserved By You"
          },
          "my_contribution": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "My Contribution"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "product_url",
          "image_url",
          "notes",
          "price",
          "mode",
          "target_amount",
          "collected_amount",
          "status",
          "position",
          "is_reserved",
          "reserved_by_you",
          "my_contribution",
          "progress_percent"
        ],
        "title": "GuestItemView"
      },
      "GuestSessionCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GuestSessionCreateRequest"
      },
      "GuestSessionResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "guest_session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Guest Session Id"
          },
          "guest_name": {
            "type": "string",
            "title": "Guest Name"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "guest_session_id",
          "guest_name",
          "expires_at"
        ],
        "title": "GuestSessionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemMode": {
        "type": "string",
        "enum": [
          "single",
          "group"
        ],
        "title": "ItemMode"
      },
...
      "OwnerItemView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "product_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "mode": {
            "$ref": "#/components/schemas/ItemMode"
          },
          "target_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Amount"
          },
          "collected_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Collected Amount"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "is_reserved": {
            "type": "boolean",
            "title": "Is Reserved"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "product_url",
          "image_url",
          "notes",
          "price",
          "mode",
          "target_amount",
          "collected_amount",
          "status",
          "position",
          "is_reserved",
          "progress_percent"
        ],
        "title": "OwnerItemView"
      },
...
      "PublicWishlistView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "occasion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occasion"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "status": {
            "$ref": "#/components/schemas/WishlistStatus"
          },
          "share_slug": {
            "type": "string",
            "title": "Share Slug"
          },
          "viewer_kind": {
            "type": "string",
            "enum": [
              "anonymous",
              "guest",
              "owner"
            ],
            "title": "Viewer Kind"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GuestItemView"
                },
                "type": "array"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/OwnerItemView"
                },
                "type": "array"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "occasion",
          "currency",
          "event_date",
          "status",
          "share_slug",
          "viewer_kind",
          "items"
        ],
        "title": "PublicWishlistView"
      },
...
      "ReservationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "is_reserved": {
            "type": "boolean",
            "title": "Is Reserved"
          }
        },
        "type": "object",
        "required": [
          "message",
          "item_id",
          "is_reserved"
        ],
        "title": "ReservationResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WishlistStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "closed"
        ],
        "title": "WishlistStatus"
      },
...
    }
  }
}